Move playSound code to utils
